Break out lighthouse command
